requirements.txt
# IBM watsonx SDK
ibm-watsonx-ai>=1.1.0

# Vector database for RAG
chromadb>=0.4.0

# Web framework
streamlit>=1.30.0

# Document processing
langchain>=0.1.0
langchain-community>=0.0.20

# Text processing
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
